feat(tasks): secure tasks endpoints for current user

+++ src/tasks/tasks.controller.ts
@@ -3,6 +3,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -22,6 +23,7 @@ import { CreateTaskLabelDto } from './create-task-label.dto';
 import { FindTaskParams } from './find-task.params';
 import { PaginationParams } from '../common/pagination.params';
 import { PaginationResponse } from '../common/pagination.response';
+import { CurrentUserId } from '../users/decorators/current-user-id.decorator';
 
 @Controller('tasks')
 export class TasksController {
@@ -45,21 +47,36 @@ export class TasksController {
   }
 
   @Get('/:id')
-  public async findOne(@Param() params: FindOneParams): Promise<Task> {
-    return await this.findOneOrFail(params.id);
+  public async findOne(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+    return task;
   }
 
   @Post()
-  public async create(@Body() createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.tasksService.createTask(createTaskDto);
+  public async create(
+    @Body() createTaskDto: CreateTaskDto,
+    // @Request() request: AuthRequest,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    return await this.tasksService.createTask({
+      ...createTaskDto,
+      userId,
+    });
   }
 
   @Patch('/:id')
   public async updateTask(
     @Param() params: FindOneParams,
     @Body() updateTaskDto: UpdateTaskDto,
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+
     try {
       return await this.tasksService.updateTask(task, updateTaskDto);
     } catch (e) {
@@ -72,8 +89,13 @@ export class TasksController {
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  public async deleteTask(@Param() params: FindOneParams): Promise<void> {
+  public async deleteTask(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<void> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+
     await this.tasksService.deleteTask(task);
   }
 
@@ -84,8 +106,11 @@ export class TasksController {
   public async addLabels(
     @Param() params: FindOneParams,
     @Body() labels: CreateTaskLabelDto[],
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+
     return await this.tasksService.addLabels(task, labels);
   }
 
@@ -94,8 +119,10 @@ export class TasksController {
   async removeLabels(
     @Param() { id }: FindOneParams,
     @Body() labelNames: string[],
+    @CurrentUserId() userId: string,
   ): Promise<void> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.removeLabels(task, labelNames);
   }
 
@@ -107,5 +134,11 @@ export class TasksController {
     }
 
     return task;
+  }
+
+  private checkTaskOwnership(task: Task, userId: string): void {
+    if (task.userId !== userId) {
+      throw new ForbiddenException('You can only accept your own tasks');
+    }
   }
 }
